Overhauled file organisation. Read INSTALLATION.md for new rules
