+++ .gitignore
@@ -1,3 +1,15 @@
+# Ignore build directory
+build/
+
+# Ignore binary output directory
+bin/
+
+# Ignore MacOS system files
+.DS_Store
+
+# Ignore linux system files
+*~
+
 # Prerequisites
 *.d
 
@@ -50,5 +62,3 @@ modules.order
 Module.symvers
 Mkfile.old
 dkms.conf
-.vscode/
-bin/
