+++ src/main.c
@@ -1,0 +1,6 @@
+#include "../include/shell.h"
+
+int main(int argc, char **argv) {
+  cpsh_loop();  // Call the cpsh_loop function
+  return 0;
+}
